WIP: Fairy Helper em debug + fork autossuficiente (sessao 2026-05-26)
Trabalho em andamento (NAO validado ao vivo ainda):
- Fairy reescrita LIMPA (modo Helper: cura 2 -> espera -> P seguir); ordem 2-entao-P.
- buff/pet viraram STUBS 'em reconstrucao' (logica e abas).
- Removido target_self (F1, codigo morto) + submodulo externo talisman_online_python
  + remote 'upstream' + URLs do chestm007 (pyproject/discord/README) -> fork autossuficiente.
- DEBUG: instrumentacao TEMPORARIA no client_window.left_click (loga cada clique) -- REMOVER depois.

CAUSA RAIZ do auto-select (a consertar na volta): o bot NAO seleciona o aliado
antes de curar -> cura(2) sem alvo -> o JOGO mira em si mesmo. Provado que o bot
nao clica em (40,40) (era servidor velho na memoria). Falta: clicar no aliado
(retrato do grupo team_1, nao o player=(40,40)) antes da cura. Detalhe na memoria.

+++ src/GhostBot/config.py
@@ -162,7 +162,7 @@ class FairyConfig(FunctionConfig):
     buff_self: bool = None
     # Modo Helper (cross-PC): cura + segue (tecla P, default) um aliado SO por tecla.
     helper_mode: bool = None
-    heal_interval_secs: int = None   # "cura sempre" a cada X segundos
+    heal_interval_secs: float = None   # "cura sempre" a cada X segundos (decimal, ex: 2.5)
 
 @dataclass
 class SellConfig(FunctionConfig):
